refactor(all-clients): extract common client provider validation to a common validator
